interpreterをファイル分割
